+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Added `\Wizaplace\SDK\Company\CompanyService::updateFile`
 - Added `\Wizaplace\SDK\Company\CompanyService::deleteFile`
+- Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationFromUserId`
+- Added `\Wizaplace\SDK\Organisation\OrganisationService::addUserToGroup`
+- Added `\Wizaplace\SDK\Organisation\OrganisationService::removeUserToGroup`
+- Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationBaskets`
 
 ### 1.33.0